Struct wiremock::Request

pub struct Request {
    pub url: Url,
    pub method: Method,
    pub headers: HashMap<HeaderName, HeaderValues>,
    pub body: Vec<u8>,
    pub body_print_limit: BodyPrintLimit,
}

An incoming request to an instance of MockServer.

Each matcher gets an immutable reference to a Request instance in the matches method defined in the Match trait.

Implementation notes:

We can’t use http_types::Request directly in our Match::matches signature: it requires having mutable access to the request to extract the body (which gets consumed when read!). It would also require matches to be async, which is cumbersome due to the lack of async traits.

We introduce our Request type to perform this extraction once when the request arrives in the mock serve, store the result and pass an immutable reference to it to all our matchers.

Fields
url: Urlmethod: Methodheaders: HashMap<HeaderName, HeaderValues>body: Vec<u8>body_print_limit: BodyPrintLimit
Implementations

impl Request

pub fn body_json<T: DeserializeOwned>(&self) -> Result<T, Error>

pub async fn from(request: Request) -> Request